third_party/blink/web_tests/external/wpt/permissions-policy/idlharness.window-expected.txt

This is a testharness.js-based test.
Found 34 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] PermissionsPolicy interface: existence and properties of interface object
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy interface object length
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy interface object name
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy interface: operation allowsFeature(DOMString, optional DOMString)
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy interface: operation features()
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy interface: operation allowedFeatures()
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy interface: operation getAllowlistForFeature(DOMString)
  assert_own_property: self does not have own property "PermissionsPolicy" expected property "PermissionsPolicy" missing
[FAIL] PermissionsPolicy must be primary interface of document.permissionsPolicy
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] Stringification of document.permissionsPolicy
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PermissionsPolicy interface: document.permissionsPolicy must inherit property "allowsFeature(DOMString, optional DOMString)" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PermissionsPolicy interface: calling allowsFeature(DOMString, optional DOMString) on document.permissionsPolicy with too few arguments must throw TypeError
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PermissionsPolicy interface: document.permissionsPolicy must inherit property "features()" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PermissionsPolicy interface: document.permissionsPolicy must inherit property "allowedFeatures()" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PermissionsPolicy interface: document.permissionsPolicy must inherit property "getAllowlistForFeature(DOMString)" with the proper type
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PermissionsPolicy interface: calling getAllowlistForFeature(DOMString) on document.permissionsPolicy with too few arguments must throw TypeError
  assert_equals: wrong typeof object expected "object" but got "undefined"
[FAIL] PermissionsPolicyViolationReportBody interface: existence and properties of interface object
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface object length
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface object name
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface: operation toJSON()
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface: attribute featureId
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface: attribute sourceFile
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface: attribute lineNumber
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface: attribute columnNumber
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] PermissionsPolicyViolationReportBody interface: attribute disposition
  assert_own_property: self does not have own property "PermissionsPolicyViolationReportBody" expected property "PermissionsPolicyViolationReportBody" missing
[FAIL] HTMLIFrameElement interface: attribute permissionsPolicy
  assert_true: The prototype object must have a property "permissionsPolicy" expected true got false
[FAIL] HTMLIFrameElement interface: document.createElement("iframe") must inherit property "permissionsPolicy" with the proper type
  assert_inherits: property "permissionsPolicy" not found in prototype chain
[FAIL] Document interface: attribute permissionsPolicy
  assert_true: The prototype object must have a property "permissionsPolicy" expected true got false
[FAIL] Document interface: document must inherit property "permissionsPolicy" with the proper type
  assert_inherits: property "permissionsPolicy" not found in prototype chain
Harness: the test ran to completion.
